💥  @jeromefitz/conventional-gitmoji (init) 🎉️
♻️  conventional-gitmoji (hack)

This *cough* needs to go first.

🔥  unused script

💥️  BREAKING CHANGE: conventional-gitmoji
