+++ packages/codestyle/tsconfig.json
@@ -2,9 +2,9 @@
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "@jeromefitz/codestyle",
   "compilerOptions": {
-    "lib": ["ES2020"],
-    "module": "ES2020",
-    "target": "ES2020",
+    "lib": ["esnext"],
+    "module": "esnext",
+    "target": "esnext",
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
     "downlevelIteration": true,
