+++ packages/codestyle/tsconfig.node.json
@@ -2,9 +2,9 @@
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "@jeromefitz/codestyle/node",
   "compilerOptions": {
-    "lib": ["ES2020"],
-    "module": "ES2020",
-    "target": "ES2020",
+    "lib": ["esnext"],
+    "module": "esnext",
+    "target": "esnext",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
