+++ packages/conventional-gitmoji/src/index.cjs
@@ -1,0 +1,5 @@
+const releaseRules = require('./releaseRules.cjs')
+const types = require('./types.cjs')
+const typeSpecs = require('./typeSpecs.cjs')
+
+module.exports = { releaseRules, types, typeSpecs }
