+++ packages/conventional-gitmoji/src/types.cjs
@@ -1,0 +1,869 @@
+const types = {
+  access: {
+    branch: false,
+    code: ':wheelchair:',
+    commit: 'access',
+    description: 'Improve accessibility.',
+    emoji: '♿️',
+    entity: '&#9855;',
+    name: 'wheelchair',
+    releaseNotes: true,
+    section: 'Improve accessibility.',
+    semver: 'patch',
+  },
+  analytics: {
+    branch: false,
+    code: ':chart_with_upwards_trend:',
+    commit: 'analytics',
+    description: 'Add or update analytics or track code.',
+    emoji: '📈',
+    entity: '&#x1F4C8;',
+    name: 'chart-with-upwards-trend',
+    releaseNotes: true,
+    section: 'Add or update analytics or track code.',
+    semver: 'patch',
+  },
+  animation: {
+    branch: false,
+    code: ':dizzy:',
+    commit: 'animation',
+    description: 'Add or update animations and transitions.',
+    emoji: '💫',
+    entity: '&#x1f4ab;',
+    name: 'animation',
+    releaseNotes: true,
+    section: 'Add or update animations and transitions.',
+    semver: 'patch',
+  },
+  arch: {
+    branch: false,
+    code: ':building_construction:',
+    commit: 'arch',
+    description: 'Make architectural changes.',
+    emoji: '🏗️',
+    entity: '&#1f3d7;',
+    name: 'building-construction',
+    releaseNotes: true,
+    section: 'Make architectural changes.',
+    semver: null,
+  },
+  assets: {
+    branch: false,
+    code: ':bento:',
+    commit: 'assets',
+    description: 'Add or update assets.',
+    emoji: '🍱',
+    entity: '&#1F371',
+    name: 'bento',
+    releaseNotes: true,
+    section: 'Add or update assets.',
+    semver: 'patch',
+  },
+  beer: {
+    branch: false,
+    code: ':beers:',
+    commit: 'beer',
+    description: 'Write code drunkenly.',
+    emoji: '🍻',
+    entity: '&#x1f37b;',
+    name: 'beers',
+    releaseNotes: false,
+    section: 'Write code drunkenly.',
+    semver: null,
+  },
+  breaking: {
+    branch: false,
+    code: ':boom:',
+    commit: 'breaking',
+    description: 'Introduce breaking changes.',
+    emoji: '💥',
+    entity: '&#x1f4a5;',
+    name: 'boom',
+    releaseNotes: true,
+    section: 'Introduce breaking changes.',
+    semver: 'major',
+  },
+  build: {
+    branch: false,
+    code: ':hammer:',
+    commit: 'build',
+    description: 'Add or update development scripts.',
+    emoji: '🔨',
+    entity: '&#128296;',
+    name: 'hammer',
+    releaseNotes: true,
+    section: 'Add or update development scripts.',
+    semver: null,
+  },
+  catch: {
+    branch: false,
+    code: ':goal_net:',
+    commit: 'catch',
+    description: 'Catch errors.',
+    emoji: '🥅',
+    entity: '&#x1F945;',
+    name: 'goal-net',
+    releaseNotes: true,
+    section: 'Catch errors.',
+    semver: 'patch',
+  },
+  chore: {
+    branch: 'chore',
+    code: ':computer_disk:',
+    commit: 'chore',
+    description: 'Changes that don’t modify src or test files',
+    emoji: '💽️',
+    entity: '&#x1f4bd;',
+    name: 'computer-disk',
+    releaseNotes: false,
+    section: 'Changes that don’t modify src or test files',
+    semver: null,
+  },
+  ci: {
+    branch: 'ci',
+    code: ':construction_worker:',
+    commit: 'ci',
+    description: 'Add or update CI build system.',
+    emoji: '👷',
+    entity: '&#x1f477;',
+    name: 'construction-worker',
+    releaseNotes: true,
+    section: 'Add or update CI build system.',
+    semver: null,
+  },
+  clean: {
+    branch: false,
+    code: ':wastebasket:',
+    commit: 'clean',
+    description: 'Deprecate code that needs to be cleaned up.',
+    emoji: '🗑️',
+    entity: '&#x1F5D1;',
+    name: 'wastebasket',
+    releaseNotes: true,
+    section: 'Deprecate code that needs to be cleaned up.',
+    semver: 'patch',
+  },
+  compat: {
+    branch: false,
+    code: ':alien:',
+    commit: 'compat',
+    description: 'Update code due to external API changes.',
+    emoji: '👽️',
+    entity: '&#1F47D;',
+    name: 'alien',
+    releaseNotes: true,
+    section: 'Update code due to external API changes.',
+    semver: 'patch',
+  },
+  config: {
+    branch: false,
+    code: ':wrench:',
+    commit: 'config',
+    description: 'Add or update configuration files.',
+    emoji: '🔧',
+    entity: '&#x1f527;',
+    name: 'wrench',
+    releaseNotes: true,
+    section: 'Add or update configuration files.',
+    semver: 'patch',
+  },
+  'contrib-add': {
+    branch: false,
+    code: ':busts_in_silhouette:',
+    commit: 'contrib-add',
+    description: 'Add or update contributor(s).',
+    emoji: '👥',
+    entity: '&#128101;',
+    name: 'busts-in-silhouette',
+    releaseNotes: true,
+    section: 'Add or update contributor(s).',
+    semver: null,
+  },
+  data: {
+    branch: false,
+    code: ':monocle_face:',
+    commit: 'data',
+    description: 'Data exploration/inspection.',
+    emoji: '🧐',
+    entity: '&#x1F9D0;',
+    name: 'monocle-face',
+    releaseNotes: true,
+    section: 'Data exploration/inspection.',
+    semver: null,
+  },
+  db: {
+    branch: false,
+    code: ':card_file_box:',
+    commit: 'db',
+    description: 'Perform database related changes.',
+    emoji: '🗃️',
+    entity: '&#128451;',
+    name: 'card-file-box',
+    releaseNotes: true,
+    section: 'Perform database related changes.',
+    semver: 'patch',
+  },
+  'dep-add': {
+    branch: false,
+    code: ':heavy_plus_sign:',
+    commit: 'dep-add',
+    description: 'Add a dependency.',
+    emoji: '➕',
+    entity: '&#10133;',
+    name: 'heavy-plus-sign',
+    releaseNotes: true,
+    section: 'Add a dependency.',
+    semver: 'patch',
+  },
+  'dep-rm': {
+    branch: false,
+    code: ':heavy_minus_sign:',
+    commit: 'dep-rm',
+    description: 'Remove a dependency.',
+    emoji: '➖',
+    entity: '&#10134;',
+    name: 'heavy-minus-sign',
+    releaseNotes: true,
+    section: 'Remove a dependency.',
+    semver: 'patch',
+  },
+  'dep-up': {
+    branch: false,
+    code: ':package:',
+    commit: 'dep-up',
+    description: 'Add or update compiled files or packages.',
+    emoji: '📦️',
+    entity: '&#1F4E6;',
+    name: 'package',
+    releaseNotes: true,
+    section: 'Add or update compiled files or packages.',
+    semver: 'patch',
+  },
+  deploy: {
+    branch: false,
+    code: ':rocket:',
+    commit: 'deploy',
+    description: 'Deploy stuff.',
+    emoji: '🚀',
+    entity: '&#x1f680;',
+    name: 'rocket',
+    releaseNotes: true,
+    section: 'Deploy stuff.',
+    semver: null,
+  },
+  docs: {
+    branch: false,
+    code: ':memo:',
+    commit: 'docs',
+    description: 'Add or update documentation.',
+    emoji: '📝',
+    entity: '&#x1f4dd;',
+    name: 'memo',
+    releaseNotes: true,
+    section: 'Add or update documentation.',
+    semver: null,
+  },
+  'docs-code': {
+    branch: false,
+    code: ':bulb:',
+    commit: 'docs-code',
+    description: 'Add or update comments in source code.',
+    emoji: '💡',
+    entity: '&#128161;',
+    name: 'bulb',
+    releaseNotes: true,
+    section: 'Add or update comments in source code.',
+    semver: null,
+  },
+  downgrade: {
+    branch: false,
+    code: ':arrow_down:',
+    commit: 'downgrade',
+    description: 'Downgrade dependencies.',
+    emoji: '⬇️',
+    entity: '⬇️',
+    name: 'arrow-down',
+    releaseNotes: true,
+    section: 'Downgrade dependencies.',
+    semver: 'patch',
+  },
+  dx: {
+    branch: false,
+    code: ':technologist:',
+    commit: 'dx',
+    description: 'Improve developer experience',
+    emoji: '🧑‍💻',
+    entity: '&#129489;&#8205;&#128187;',
+    name: 'technologist',
+    releaseNotes: true,
+    section: 'Improve developer experience',
+    semver: null,
+  },
+  egg: {
+    branch: false,
+    code: ':egg:',
+    commit: 'egg',
+    description: 'Add or update an easter egg.',
+    emoji: '🥚',
+    entity: '&#129370;',
+    name: 'egg',
+    releaseNotes: false,
+    section: 'Add or update an easter egg.',
+    semver: 'patch',
+  },
+  experiment: {
+    branch: false,
+    code: ':alembic:',
+    commit: 'experiment',
+    description: 'Perform experiments.',
+    emoji: '⚗️',
+    entity: '&#128248;',
+    name: 'alembic',
+    releaseNotes: false,
+    section: 'Perform experiments.',
+    semver: 'patch',
+  },
+  feat: {
+    branch: 'feature',
+    code: ':sparkles:',
+    commit: 'feat',
+    description: 'Introduce new features.',
+    emoji: '✨',
+    entity: '&#x2728;',
+    name: 'sparkles',
+    releaseNotes: true,
+    section: 'Introduce new features.',
+    semver: 'minor',
+  },
+  fix: {
+    branch: 'fix',
+    code: ':bug:',
+    commit: 'fix',
+    description: 'Fix a bug.',
+    emoji: '🐛',
+    entity: '&#x1f41b;',
+    name: 'bug',
+    releaseNotes: true,
+    section: 'Fix a bug.',
+    semver: 'patch',
+  },
+  'fix-ci': {
+    branch: false,
+    code: ':green_heart:',
+    commit: 'fix-ci',
+    description: 'Fix CI Build.',
+    emoji: '💚',
+    entity: '&#x1f49a;',
+    name: 'green-heart',
+    releaseNotes: true,
+    section: 'Fix CI Build.',
+    semver: null,
+  },
+  flags: {
+    branch: false,
+    code: ':triangular_flag_on_post:',
+    commit: 'flags',
+    description: 'Add, update, or remove feature flags.',
+    emoji: '🚩',
+    entity: '&#x1F6A9;',
+    name: 'triangular-flag-on-post',
+    releaseNotes: true,
+    section: 'Add, update, or remove feature flags.',
+    semver: 'patch',
+  },
+  healthcheck: {
+    branch: false,
+    code: ':stethoscope:',
+    commit: 'healthcheck',
+    description: 'Add or update healthcheck.',
+    emoji: '🩺',
+    entity: '&#x1FA7A;',
+    name: 'stethoscope',
+    releaseNotes: true,
+    section: 'Add or update healthcheck.',
+    semver: null,
+  },
+  hotfix: {
+    branch: 'hotfix',
+    code: ':ambulance:',
+    commit: 'hotfix',
+    description: 'Critical hotfix.',
+    emoji: '🚑️',
+    entity: '&#128657;',
+    name: 'ambulance',
+    releaseNotes: true,
+    section: 'Critical hotfix.',
+    semver: 'patch',
+  },
+  i18n: {
+    branch: false,
+    code: ':globe_with_meridians:',
+    commit: 'i18n',
+    description: 'Internationalization and localization.',
+    emoji: '🌐',
+    entity: '&#127760;',
+    name: 'globe-with-meridians',
+    releaseNotes: true,
+    section: 'Internationalization and localization.',
+    semver: 'patch',
+  },
+  ignore: {
+    branch: false,
+    code: ':see_no_evil:',
+    commit: 'ignore',
+    description: 'Add or update a .gitignore file.',
+    emoji: '🙈',
+    entity: '&#8bdfe7;',
+    name: 'see-no-evil',
+    releaseNotes: false,
+    section: 'Add or update a .gitignore file.',
+    semver: null,
+  },
+  inf: {
+    branch: false,
+    code: ':bricks:',
+    commit: 'inf',
+    description: 'Infrastructure related changes.',
+    emoji: '🧱',
+    entity: '&#x1f9f1;',
+    name: 'bricks',
+    releaseNotes: true,
+    section: 'Infrastructure related changes.',
+    semver: null,
+  },
+  init: {
+    branch: false,
+    code: ':tada:',
+    commit: 'init',
+    description: 'Begin a project.',
+    emoji: '🎉',
+    entity: '&#127881;',
+    name: 'tada',
+    releaseNotes: true,
+    section: 'Begin a project.',
+    semver: null,
+  },
+  iphone: {
+    branch: false,
+    code: ':iphone:',
+    commit: 'iphone',
+    description: 'Work on responsive design.',
+    emoji: '📱',
+    entity: '&#128241;',
+    name: 'iphone',
+    releaseNotes: true,
+    section: 'Work on responsive design.',
+    semver: 'patch',
+  },
+  license: {
+    branch: false,
+    code: ':page_facing_up:',
+    commit: 'license',
+    description: 'Add or update license.',
+    emoji: '📄',
+    entity: '&#1F4C4;',
+    name: 'page-facing-up',
+    releaseNotes: true,
+    section: 'Add or update license.',
+    semver: null,
+  },
+  lint: {
+    branch: false,
+    code: ':rotating_light:',
+    commit: 'lint',
+    description: 'Fix compiler / linter warnings.',
+    emoji: '🚨',
+    entity: '&#x1f6a8;',
+    name: 'rotating-light',
+    releaseNotes: true,
+    section: 'Fix compiler / linter warnings.',
+    semver: null,
+  },
+  'log-add': {
+    branch: false,
+    code: ':loud_sound:',
+    commit: 'log-add',
+    description: 'Add or update logs.',
+    emoji: '🔊',
+    entity: '&#128266;',
+    name: 'loud-sound',
+    releaseNotes: false,
+    section: 'Add or update logs.',
+    semver: null,
+  },
+  'log-rm': {
+    branch: false,
+    code: ':mute:',
+    commit: 'log-rm',
+    description: 'Remove logs.',
+    emoji: '🔇',
+    entity: '&#128263;',
+    name: 'mute',
+    releaseNotes: false,
+    section: 'Remove logs.',
+    semver: null,
+  },
+  logic: {
+    branch: false,
+    code: ':necktie:',
+    commit: 'logic',
+    description: 'Add or update business logic',
+    emoji: '👔',
+    entity: '&#128084;',
+    name: 'necktie',
+    releaseNotes: true,
+    section: 'Add or update business logic',
+    semver: 'patch',
+  },
+  merge: {
+    branch: false,
+    code: ':twisted_rightwards_arrows:',
+    commit: 'merge',
+    description: 'Merge branches.',
+    emoji: '🔀',
+    entity: '&#128256;',
+    name: 'twisted-rightwards-arrows',
+    releaseNotes: false,
+    section: 'Merge branches.',
+    semver: null,
+  },
+  mock: {
+    branch: false,
+    code: ':clown_face:',
+    commit: 'mock',
+    description: 'Mock things.',
+    emoji: '🤡',
+    entity: '&#129313;',
+    name: 'clown-face',
+    releaseNotes: true,
+    section: 'Mock things.',
+    semver: null,
+  },
+  mv: {
+    branch: false,
+    code: ':truck:',
+    commit: 'mv',
+    description: 'Move or rename resources (e.g.: files, paths, routes).',
+    emoji: '🚚',
+    entity: '&#1F69A;',
+    name: 'truck',
+    releaseNotes: false,
+    section: 'Move or rename resources (e.g.: files, paths, routes).',
+    semver: null,
+  },
+  patch: {
+    branch: false,
+    code: ':adhesive_bandage:',
+    commit: 'patch',
+    description: 'Simple fix for a non-critical issue.',
+    emoji: '🩹',
+    entity: '&#x1FA79;',
+    name: 'adhesive-bandage',
+    releaseNotes: true,
+    section: 'Simple fix for a non-critical issue.',
+    semver: 'patch',
+  },
+  perf: {
+    branch: false,
+    code: ':zap:',
+    commit: 'perf',
+    description: 'Improve performance.',
+    emoji: '⚡️',
+    entity: '&#x26a1;',
+    name: 'zap',
+    releaseNotes: true,
+    section: 'Improve performance.',
+    semver: 'patch',
+  },
+  poo: {
+    branch: false,
+    code: ':poop:',
+    commit: 'poo',
+    description: 'Write bad code that needs to be improved.',
+    emoji: '💩',
+    entity: '&#58613;',
+    name: 'poop',
+    releaseNotes: false,
+    section: 'Write bad code that needs to be improved.',
+    semver: null,
+  },
+  prune: {
+    branch: false,
+    code: ':fire:',
+    commit: 'prune',
+    description: 'Remove code or files.',
+    emoji: '🔥',
+    entity: '&#x1f525;',
+    name: 'fire',
+    releaseNotes: true,
+    section: 'Remove code or files.',
+    semver: null,
+  },
+  pushpin: {
+    branch: false,
+    code: ':pushpin:',
+    commit: 'pushpin',
+    description: 'Pin dependencies to specific versions.',
+    emoji: '📌',
+    entity: '&#x1F4CC;',
+    name: 'pushpin',
+    releaseNotes: true,
+    section: 'Pin dependencies to specific versions.',
+    semver: 'patch',
+  },
+  refactor: {
+    branch: 'refactor',
+    code: ':recycle:',
+    commit: 'refactor',
+    description: 'Refactor code.',
+    emoji: '♻️',
+    entity: '&#x2672;',
+    name: 'recycle',
+    releaseNotes: true,
+    section: 'Refactor code.',
+    semver: null,
+  },
+  release: {
+    branch: 'release',
+    code: ':bookmark:',
+    commit: 'release',
+    description: 'Release / Version tags.',
+    emoji: '🔖',
+    entity: '&#x1f516;',
+    name: 'bookmark',
+    releaseNotes: false,
+    section: 'Release / Version tags.',
+    semver: null,
+  },
+  revert: {
+    branch: false,
+    code: ':rewind:',
+    commit: 'revert',
+    description: 'Revert changes.',
+    emoji: '⏪️',
+    entity: '&#9194;',
+    name: 'rewind',
+    releaseNotes: false,
+    section: 'Revert changes.',
+    semver: 'patch',
+  },
+  rip: {
+    branch: false,
+    code: ':coffin:',
+    commit: 'rip',
+    description: 'Remove dead code.',
+    emoji: '⚰️',
+    entity: '&#x26B0;',
+    name: 'coffin',
+    releaseNotes: false,
+    section: 'Remove dead code.',
+    semver: null,
+  },
+  roles: {
+    branch: false,
+    code: ':passport_control:',
+    commit: 'roles',
+    description: 'Work on code related to authorization, roles and permissions.',
+    emoji: '🛂',
+    entity: '&#x1F6C2;',
+    name: 'passport-control',
+    releaseNotes: true,
+    section: 'Work on code related to authorization, roles and permissions.',
+    semver: 'patch',
+  },
+  rollforward: {
+    branch: false,
+    code: ':fast_forward:',
+    commit: 'rollforward',
+    description: 'Create rollforward version.',
+    emoji: '⏩️',
+    entity: '&#23E9;',
+    name: 'rollforward',
+    releaseNotes: false,
+    section: 'Create rollforward version.',
+    semver: null,
+  },
+  'run-build': {
+    branch: false,
+    code: ':rocket:',
+    commit: 'run-build',
+    description: 'Custom type for CI/CD to hook into run build override.',
+    emoji: '🚀️',
+    entity: '&#1F680;',
+    name: 'run-build',
+    releaseNotes: false,
+    section: 'Custom type for CI/CD to hook into run build override.',
+    semver: 'patch',
+  },
+  security: {
+    branch: false,
+    code: ':lock:',
+    commit: 'security',
+    description: 'Fix security issues.',
+    emoji: '🔒️',
+    entity: '&#x1f512;',
+    name: 'lock',
+    releaseNotes: true,
+    section: 'Fix security issues.',
+    semver: 'patch',
+  },
+  seed: {
+    branch: false,
+    code: ':seedling:',
+    commit: 'seed',
+    description: 'Add or update seed files.',
+    emoji: '🌱',
+    entity: '&#127793;',
+    name: 'seedling',
+    releaseNotes: true,
+    section: 'Add or update seed files.',
+    semver: null,
+  },
+  seo: {
+    branch: false,
+    code: ':mag:',
+    commit: 'seo',
+    description: 'Improve SEO.',
+    emoji: '🔍️',
+    entity: '&#128269;',
+    name: 'mag',
+    releaseNotes: true,
+    section: 'Improve SEO.',
+    semver: 'patch',
+  },
+  snapshot: {
+    branch: false,
+    code: ':camera_flash:',
+    commit: 'snapshot',
+    description: 'Add or update snapshots.',
+    emoji: '📸',
+    entity: '&#128248;',
+    name: 'camera-flash',
+    releaseNotes: false,
+    section: 'Add or update snapshots.',
+    semver: null,
+  },
+  style: {
+    branch: false,
+    code: ':art:',
+    commit: 'style',
+    description: 'Improve structure / format of the code.',
+    emoji: '🎨',
+    entity: '&#x1f3a8;',
+    name: 'art',
+    releaseNotes: true,
+    section: 'Improve structure / format of the code.',
+    semver: null,
+  },
+  test: {
+    branch: 'test',
+    code: ':white_check_mark:',
+    commit: 'test',
+    description: 'Add, update, or pass tests.',
+    emoji: '✅',
+    entity: '&#x2705;',
+    name: 'white-check-mark',
+    releaseNotes: true,
+    section: 'Add, update, or pass tests.',
+    semver: null,
+  },
+  'test-fail': {
+    branch: false,
+    code: ':test_tube:',
+    commit: 'test-fail',
+    description: 'Add a failing test.',
+    emoji: '🧪',
+    entity: '&#x1F9EA;',
+    name: 'test-tube',
+    releaseNotes: true,
+    section: 'Add a failing test.',
+    semver: null,
+  },
+  texts: {
+    branch: false,
+    code: ':speech_balloon:',
+    commit: 'texts',
+    description: 'Add or update text and literals.',
+    emoji: '💬',
+    entity: '&#128172;',
+    name: 'speech-balloon',
+    releaseNotes: true,
+    section: 'Add or update text and literals.',
+    semver: 'patch',
+  },
+  types: {
+    branch: false,
+    code: ':label:',
+    commit: 'types',
+    description: 'Add or update types.',
+    emoji: '🏷️',
+    entity: '&#127991;',
+    name: 'label',
+    releaseNotes: true,
+    section: 'Add or update types.',
+    semver: 'patch',
+  },
+  typo: {
+    branch: false,
+    code: ':pencil2:',
+    commit: 'typo',
+    description: 'Fix typos.',
+    emoji: '✏️',
+    entity: '&#59161;',
+    name: 'pencil2',
+    releaseNotes: false,
+    section: 'Fix typos.',
+    semver: 'patch',
+  },
+  ui: {
+    branch: false,
+    code: ':lipstick:',
+    commit: 'ui',
+    description: 'Add or update the UI and style files.',
+    emoji: '💄',
+    entity: '&#ff99cc;',
+    name: 'lipstick',
+    releaseNotes: true,
+    section: 'Add or update the UI and style files.',
+    semver: 'patch',
+  },
+  upgrade: {
+    branch: false,
+    code: ':arrow_up:',
+    commit: 'upgrade',
+    description: 'Upgrade dependencies.',
+    emoji: '⬆️',
+    entity: '⬆️',
+    name: 'arrow-up',
+    releaseNotes: true,
+    section: 'Upgrade dependencies.',
+    semver: 'patch',
+  },
+  ux: {
+    branch: false,
+    code: ':children_crossing:',
+    commit: 'ux',
+    description: 'Improve user experience / usability.',
+    emoji: '🚸',
+    entity: '&#128696;',
+    name: 'children-crossing',
+    releaseNotes: true,
+    section: 'Improve user experience / usability.',
+    semver: 'patch',
+  },
+  wip: {
+    branch: false,
+    code: ':construction:',
+    commit: 'wip',
+    description: 'Work in progress.',
+    emoji: '🚧',
+    entity: '&#x1f6a7;',
+    name: 'construction',
+    releaseNotes: false,
+    section: 'Work in progress.',
+    semver: null,
+  },
+}
+
+// export default types
+module.exports = types
