+++ packages/git-cz/package.json
@@ -42,7 +42,6 @@
     "build": "tsc",
     "clean": "rimraf dist",
     "dry:run": "yarn clean && tsc && node ./dist/cli",
-    "gitmoji": "node ./scripts/gitmoji/index.js",
     "semantic-release": "semantic-release",
     "test:coverage": "echo 🚧️ TODO jest --coverage",
     "test:dev": "echo 🚧️ TODO jest --watch",
@@ -52,17 +51,16 @@
   "devDependencies": {
     "chai": "4.3.4",
     "commitizen": "4.2.4",
-    "isomorphic-unfetch": "3.1.0",
     "jest": "27.4.3",
     "mocha": "9.1.3",
     "rimraf": "3.0.2",
     "spawncommand": "2.2.0"
   },
   "dependencies": {
+    "@jeromefitz/conventional-gitmoji": "*",
     "@types/chalk-pipe": "2.0.0",
     "chalk": "5.0.0",
     "chalk-pipe": "5.1.1",
-    "fast-json-stable-stringify": "2.1.0",
     "fuzzy": "0.1.3",
     "inquirer": "8.2.0",
     "inquirer-list-search-prompt": "1.0.2",
