--- packages/git-cz/scripts/gitmoji/index.js
@@ -1,11 +1,0 @@
-import gitmojiBuild from './gitmojiBuild.js'
-import gitmojiFetch from './gitmojiFetch.js'
-import gitmojiTheme from './gitmojiTheme.js'
-
-const init = async () => {
-  const { gitmojis } = await gitmojiFetch()
-  const types = await gitmojiBuild(gitmojis)
-  gitmojiTheme(types)
-}
-
-void init()
