+++ packages/semantic/src/config/release-branch-types/refactor.cjs
@@ -1,3 +1,3 @@
-const refactor = []
+const refactor = ['gitmojis']
 
 module.exports = refactor
